requirements.txt
fastapi>=0.110,<0.112
uvicorn[standard]>=0.29,<0.31
pydantic>=1.10,<1.11
numpy>=1.24,<2.5
spiceypy>=6.0,<7.0
pyerfa>=2.0,<2.1
httpx>=0.27,<0.28
pytest>=8.0,<9.0
